# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Science Compiler API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /papers:
    get:
      operationId: listPapers
      tags: [papers]
      summary: List papers with optional filtering
      parameters:
        - name: topicId
          in: query
          schema:
            type: ["integer", "null"]
        - name: methodologyType
          in: query
          schema:
            type: ["string", "null"]
        - name: evidenceQuality
          in: query
          schema:
            type: ["string", "null"]
        - name: replicationStatus
          in: query
          schema:
            type: ["string", "null"]
        - name: domain
          in: query
          description: Filter papers by topic domain (e.g. "Endocrinology")
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of papers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PapersListResponse"
    post:
      operationId: createPaper
      tags: [papers]
      summary: Create a new paper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaperBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paper"
  /papers/{id}:
    get:
      operationId: getPaper
      tags: [papers]
      summary: Get paper by ID with its claims
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Paper with claims
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePaper
      tags: [papers]
      summary: Update a paper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaperBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paper"
    delete:
      operationId: deletePaper
      tags: [papers]
      summary: Delete a paper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Paper:
      type: object
      properties:
        id:
          type: integer
        topicId:
          type: integer
        title:
          type: string
        authors:
          type: string
        journal:
          type: string
        publicationYear:
          type: integer
        doi:
          type: ["string", "null"]
        pmid:
          type: ["string", "null"]
        abstract:
          type: string
        methodologyType:
          type: string
        sampleSize:
          type: ["integer", "null"]
        pValue:
          type: ["string", "null"]
        evidenceQuality:
          type: string
        replicationStatus:
          type: string
        openAccessUrl:
          type: ["string", "null"]
        isPreprint:
          type: integer
        fullTextStatus:
          type: string
        fullTextSource:
          type: ["string", "null"]
        fullTextUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - topicId
        - title
        - authors
        - journal
        - publicationYear
        - abstract
        - methodologyType
        - evidenceQuality
        - replicationStatus
        - createdAt
...
    PaperDetail:
      type: object
      properties:
        id:
          type: integer
        topicId:
          type: integer
        topicName:
          type: string
        title:
          type: string
        authors:
          type: string
        journal:
          type: string
        publicationYear:
          type: integer
        doi:
          type: ["string", "null"]
        pmid:
          type: ["string", "null"]
        abstract:
          type: string
        methodologyType:
          type: string
        sampleSize:
          type: ["integer", "null"]
        pValue:
          type: ["string", "null"]
        evidenceQuality:
          type: string
        replicationStatus:
          type: string
        openAccessUrl:
          type: ["string", "null"]
        isPreprint:
          type: integer
        fullTextStatus:
          type: string
        fullTextSource:
          type: ["string", "null"]
        fullTextUrl:
          type: ["string", "null"]
        sources:
          type: array
          items:
            $ref: "#/components/schemas/PaperSourceBadge"
        claims:
          type: array
          items:
            $ref: "#/components/schemas/ClaimSummary"
        createdAt:
          type: string
      required:
        - id
        - topicId
        - topicName
        - title
        - authors
        - journal
        - publicationYear
        - abstract
        - methodologyType
        - evidenceQuality
        - replicationStatus
        - sources
        - claims
        - createdAt
    PaperSourceBadge:
      type: object
      properties:
        sourceId:
          type: string
        displayName:
          type: string
        nativeId:
          type: string
        url:
          type: ["string", "null"]
      required: [sourceId, displayName, nativeId]
    PapersListResponse:
      type: object
      properties:
        papers:
          type: array
          items:
            $ref: "#/components/schemas/Paper"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - papers
        - total
        - limit
        - offset
    ClaimSummary:
      type: object
      properties:
        id:
          type: integer
        claimText:
          type: string
        direction:
          type: string
        evidenceQuality:
          type: string
        consensusStatus:
          type: ["string", "null"]
        supportingCount:
          type: ["integer", "null"]
        contradictingCount:
          type: ["integer", "null"]
        topicName:
          type: ["string", "null"]
      required:
        - id
        - claimText
        - direction
        - evidenceQuality
...
    CreatePaperBody:
      type: object
      properties:
        topicId:
          type: integer
        title:
          type: string
        authors:
          type: string
        journal:
          type: string
        publicationYear:
          type: integer
        doi:
          type: ["string", "null"]
        pmid:
          type: ["string", "null"]
        abstract:
          type: string
        methodologyType:
          type: string
        sampleSize:
          type: ["integer", "null"]
        pValue:
          type: ["string", "null"]
        evidenceQuality:
          type: string
        replicationStatus:
          type: ["string", "null"]
        openAccessUrl:
          type: ["string", "null"]
      required: [topicId, title, authors, journal, publicationYear, abstract, methodologyType, evidenceQuality]
    UpdatePaperBody:
      type: object
      properties:
        topicId:
          type: ["integer", "null"]
        title:
          type: ["string", "null"]
        authors:
          type: ["string", "null"]
        journal:
          type: ["string", "null"]
        publicationYear:
          type: ["integer", "null"]
        doi:
          type: ["string", "null"]
        pmid:
          type: ["string", "null"]
        abstract:
          type: ["string", "null"]
        methodologyType:
          type: ["string", "null"]
        sampleSize:
          type: ["integer", "null"]
        pValue:
          type: ["string", "null"]
        evidenceQuality:
          type: ["string", "null"]
        replicationStatus:
          type: ["string", "null"]
        openAccessUrl:
          type: ["string", "null"]
...