# lib/api-zod/src/generated/types/paperSourceBadge.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Science Compiler API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaperSourceBadge {
  sourceId: string;
  displayName: string;
  nativeId: string;
  /** @nullable */
  url?: string | null;
}